+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **1.2.0-dev**
 
+- Fix cycle detection with implicit chains
 - BC: Implicitly open new dependency chains with explicitly passed config values
 
 **1.1.0**